[DI] rename ResolveDefinitionTemplatesPass to ResolveChildDefinitionsPass
