+++ src/Symfony/Component/DependencyInjection/Compiler/PassConfig.php
@@ -48,7 +48,7 @@ class PassConfig
 
         $this->optimizationPasses = array(array(
             new ExtensionCompilerPass(),
-            new ResolveDefinitionTemplatesPass(),
+            new ResolveChildDefinitionsPass(),
             new ServiceLocatorTagPass(),
             new DecoratorServicePass(),
             new ResolveParameterPlaceHoldersPass(false),
